Refactored the code and added getall lambda
