--- cdk-ts-demo/lib/cdk-ts-demo-prop.ts
@@ -1,16 +1,0 @@
-import { Duration } from "aws-cdk-lib"
-import { HttpMethod } from "aws-cdk-lib/aws-events"
-
-export interface DemoApiProps {
-    lambdaFunctions: LambdaListType[]
-}
-
-export type LambdaListType = {
-    fuctionName: string,
-    fuctionId: string,
-    IntegrationId: string,
-    handler: string,
-    path: string,
-    method: HttpMethod,
-    timeOut: Duration     
-}
