+++ cdk-ts-demo/lib/cdkts-demo-constructs.ts
@@ -3,8 +3,21 @@ import * as lambda from 'aws-cdk-lib/aws-lambda';
 import * as apigateway from 'aws-cdk-lib/aws-apigateway';
 import { Duration } from "aws-cdk-lib";
 import { Construct } from "constructs"
-import { DemoApiProps, LambdaListType } from "./cdk-ts-demo-prop";
 import { Code, Runtime } from 'aws-cdk-lib/aws-lambda';
+import { HttpMethod } from 'aws-cdk-lib/aws-events';
+
+export interface DemoApiProps {
+    lambdaFunctions: LambdaListType[]
+}
+
+export type LambdaListType = {
+    fuctionName: string,
+    fuctionId: string,
+    handler: string,
+    path?: string,
+    method: HttpMethod,
+    timeOut: Duration     
+}
 
 
 export class DemoApiConstructs  extends Construct {
@@ -14,6 +27,7 @@ export class DemoApiConstructs  extends Construct {
         const lambdaList:LambdaListType[] =  prop.lambdaFunctions;
 
         const api = new apigateway.RestApi(this,'demoApi');
+        const resource = api.root.addResource('demo');
 
         lambdaList.map(value => {
             const lambdaFunction = this.createLambdaFunction(
@@ -23,11 +37,14 @@ export class DemoApiConstructs  extends Construct {
                 value.timeOut ? value.timeOut : Duration.seconds(3)
                 );
 
-                const integration = new apigateway.LambdaIntegration(lambdaFunction);
-
-                const resource = api.root.addResource('demo');
-                resource.addMethod(value.method, integration);
-
+                const integration = new apigateway.LambdaIntegration(lambdaFunction); 
+                if(value.path!= null){
+                    const pathPartResource = resource.addResource(value.path);
+                    pathPartResource.addMethod(value.method, integration);
+                }else{
+                    resource.addMethod(value.method, integration);    
+                } 
+                
         });
     }
 
@@ -36,7 +53,7 @@ export class DemoApiConstructs  extends Construct {
         return new lambda.Function(this, id , {
             functionName: name,
             runtime: Runtime.NODEJS_18_X,
-            code: Code.fromAsset('lambda'),
+            code: Code.fromAsset('src/lambda'),
             handler: handlerPath
         });
     }
